Chore: add utils for rule tests
